Update README.md
Number Guessing Game in Java
This Java code implements a number guessing game. Upon running the program, the user is presented with a menu that offers three options: playing the game, viewing the score board, or exiting the game.

Play the Game: If the user selects this option, they are prompted to enter the range of numbers they want to play with. A random number within that range is generated, and the user is then asked to guess the number. The program provides feedback (higher or lower) based on the user's guess. Once the user guesses the correct number, the number of tries is displayed, and the score is added to the score board. The menu is then displayed again.

Score Board: If the user selects this option, the program displays the score board, showing the number of tries for each completed game. The scores are sorted in ascending order. The menu is then displayed again.

Exit the Game: If the user selects this option, the program terminates, and a goodbye message is displayed.

The code uses the 'java.util.Scanner' class to handle user input and the ,java.util.ArrayList' class to store the scores achieved in each game.